tls: Implement the RFC 7919 Supported Groups extension
For ECDHE (RFC 8422) we implemented the Supported Elliptic Curves
extension which was later renamed Supported Groups in RFC 7919 and
extended to be used for negotiating FF groups for DHE too.  This change
is acknowledged in RFC 8422 too.  This fixes some problems with the DHE
key exchange in TLS although it seems like it may have been released
too late to get TLS users to re-enable DHE when ECDHE is widely
supported and had the group negotiation specified in a way free of
the issues in DHE.  RFC 7919 is only supported by few https servers
(all wikipedia servers as far as I could tell)