key: Don't generate description string from pointer
Within a keyring each key must have a unique key type + description
string combination.  If a key is linked to a keyring with the same
description and key type as another key in the keyring, the other key is
unlinked (this is documented).  Since we allow our references to the
kernel keys to be freed without revoking them, through
l_key_free_norevoke, and we depend on them staying usable in the
keyrings we can't generate the key descriptions from just the pointer
because after a key is freed in the userspace another key may get the
same pointer and the same description string.  In fact if we're adding
keys to a keyring in a loop and freeing our references to them
immediately each key will get the same description and we end up
removing the previous key in each iteration.  Fix this by using a
global (to the process) sequential ids for the keys and keyrings.