utf8: Fix expected bytes in l_utf8_get_codepoint
UTF-8 requires the form 10xxxxxx for the second, third and forth bytes
of a well-formed byte sequences. Therefore, comparing with 0 is not
sufficient to exclude ill-formed byte sequences, but the first two bit
must follow the specified form.

Without this check, iwd crashes if it encounters Latin-1 Supplement
encoded SSIDs during scanning, because they are erroneously accepted as
valid UTF-8.