dhcp: Rework DHCP Transport
The previous attempt at only using a non-raw AF_INET socket didn't
really work.  The kernel refused to deliver packets to it when the
interface was not configured, even when we requested a broadcast from
the DHCP Server, and the server honored our request.  This worked on
some kernels but not others.

So now we switch to a dual-socket approach.  We open up a RAW socket
which will be used for sending packets when we're supposed to broadcast
(e.g. in the Discover, Requesting and Rebinding phases).  This socket
will also be used for all packet reception.  A BPF filter is installed
on this socket to help filter unwanted packets and reduce wakeups.

We also setup a second AF_INET socket when the interface has been
configured.  This socket will be used to send packets that need to be
unicast and routed (e.g. in the Renewing state).