gpio: fix clang -Wpointer-bool-conversion warning
Remove a check that incorrectly attempted to determine if the "name"
field of a struct gpioline_info object was NULL.  It relied on a
pointer-to-bool conversion that didn't work as expected since the
"name" field is an array, not a pointer.  The address of an array
always evaluates to true when converted to a boolean.  In this case,
there is no need for a NULL pointer check.