tls: Drop call to l_certchain_find
The intent here was to try and build a certificate chain from a provided
certificate out to a configured CA and use that as the chain in the TLS
connection.  However, this was never actually implemented.

The call to l_certchain_find could short-circuit if a fully formed
certchain was already provided (the typical case today), with a
restriction that the certchain was verifiable by one of the provided
CAs.  This latter part was not true for networks that used different CAs
for to sign user certificates and peer certificates.