dhcp6: Build ORO according to RFC7844
Instead of manually including the 'special' options like SOL_MAX_RT and
INF_MAX_RT for Solicit and Information-Request messages, add them to a
temporary uintset.  This has the advantage that all options will be
added in-order starting with the lowest numbered option.  This is
exactly what RFC 7844 Section 4.6 recommends.