cipher: remove obsolete arc4 support
The rc4 cipher is obsolete, and is only relevant for very specific use
cases where compatibility with legacy hardware or software is needed.
Now that iwd has been updated to use its own arc4 implementation, which
is tailored to its needs, and support for the RC4 based TLS cipher suites
has been dropped, we can drop the implementation from libell.