dhcp6: Break up dhcp6_client_timeout_send
In certain circumstances it will be required to start sending the next
message immediately instead of waiting for the timeout to expire.  To
facilitate that, break up dhcp6_client_timeout_send and move logic of
sending the next message into a new function.  This also moves some code
around to avoid forward declarations.