dhcp-transport: modify BPF filter for server
In preparation for dhcp server the XID portion of the BPF
needs to be skipped when the port is DHCP_PORT_SERVER since
this value will not be present in messages going to the
server. The check for the message op was also changed to
include both BOOTREQUEST and BOOTREPLY depending if the
port matches the server or client port.