dhcp-transport: rework 'broadcast' op to allow unicast
Functionally this was being used as a broadcast only because the
destination address was a broadcast address. But providing a
MAC address would effectively turn this into a unicast over l2
which DHCP server requires.

To support server a destination MAC address was added to the
arguments and if NULL the destination will be set to a broadcast
address. The name of the op was changed to 'l2_send' to better
describe the new behavior. All client code will remain nearly the
same (just name changes) except having to set the destination MAC
to NULL indicating it is a broadcast.