dhcp-server: add timer for tracking expired leases
Currently dhcp-server has a limitation where it only calls the
lease expired callback if a client actually releases the lease
itself.

This patchs adds a single timer to the server object which tracks
the soonest expiring lease. When this lease expires the event is
triggered and the timeout is reset to the (new) soonest expiring
lease.