cipher: Add an RC2 implementation
RC2 is still used by default by tools creating PKCS#12 files (.pfx /
.p12).  Since it's not available from the kernel crypto API, add a local
implementation.  This is currently the only cipher implemented locally
so checks are needed in all our l_cipher ops to see whether a cipher is
handled locally or by the kernel.

Adding this cipher is not a recommendation of using it, it's discouraged
where it can be avoided.