pkcs5: Rename to cert-crypto
After recent changes pkcs5.c contained cryptographic routines not related
to PKCS#5 because there are ciphers in other standards defined in
similar ways.  From the pkcs5_cipher_from_alg_id() client's point of view
it didn't make sense to call different functions in different files
depending on which standard the Algorithm ID was expected or where any
well known cipher should be accepted.

cert.c and cert-crypto.c will now contain the classes and utilites related
to both certificates and private keys and low-level cryptographics
routines (in cert-crypto.c) related to the public-key cryptography
standard.  PKCS (Public-Key Cryptography Standards) would be a fitting
name but use the name "cert" to not imply that the contents are limited
to the standards created by RSA.

In this patch I rename both the file and the functions.  This is
backwards incompatible and users of the functions need to be updated.