cert: Add l_cert_load_container_file
Add a function that takes a file path and detects what X.509 certificate
and/or private key format it uses and parses it accordingly.  This is to
make it easier for clients to support multiple file formats, including
raw X.509 certificates -- without this they would have to load the
contents of the file and call l_cert_new_from_der().
l_cert_load_container can also be used instead of
l_pem_load_certificate_chain() and l_pem_load_private_key().

This function can now load binary certificate files which are not PEM,
that's why it's placed in cert.c but a bunch of utilities in pem.c are
exposed in pem-private.h for this function.. It treats PEM files as
containers that can have both private keys and certificates at the same
time, openssl can parse such files and also produces such files in some
situations and they may have some good uses.