tester: Add framework for serialized test execution
This adds a framework for running series of noninteractive tests
that are capable of executing more complex scenarios than a usual unit
test.

The framework enables staging of a test case processing:
- test pre-setup: can be used to prepare an initial testing environment
                  potentially common to between several test cases, e.g.,
		  put an emulator in a certain well known state.
- test setup: used to setup conditions specific for the test case
- test run: execute actual test case
- test teardown: Cleans up after test setup
- test post-teardown: Cleans up after test pre-setup

Additional command line options allow to run a selected group of tests
based either on a group prefix of the test name or a substring in test
description.