dhcp-server: Use stricter max timeout check
The next_timeout_ms value is sanity checked against ULONG_MAX to make
sure that we do not set a timeout that can overflow.  However, this is
not a good value to check against since it is different depending on
32/64 bit platforms.  Also, timerfd is limited to 32-bit second value on
most systems, so the better check would have been to compare against
UINT32_MAX * L_SEC_TO_MSEC...

We can further narrow down the max timeout value since we have a single
lease timeout setting.  So we use that instead and print a warning if
the timeout somehow exceeds that value.