dhcp: Add logging if bind function fails
The bind function returns any errors it encounters but they were
silently being discarded in this method. This at least logs this if
debug messages are enabled.

In the future it would be nice to have this log even if the debug output
isn't enabled as DHCP will not be working correctly.