dhcp-server: Refactor lease lookup
A few places in the code assume that there can be only one lease per
client MAC address (or recently client ID).  The spec doesn't actually
say this and says that a lease is identified by the client-id/mac plus
the IP:

"The combination of 'client identifier' or 'chaddr' and assigned network
address constitute a unique identifier for the client's lease and are
used by both the client and server to identify a lease referred to in
any DHCP messages."

Start identifying leases this way.  Don't actually add any way for a
client to obtain a second lease -- we still ignore the 'requested IP'
option in DHCPDISCOVER if the client already has a lease.

This currently solves an unrelated problem: there was no upper bound
on server->expired_list size.  When handling DHCPDISCOVER we'd free the
oldest lease in server->expired_list only under some conditions,
otherwise we might have been allocating an IP that was already in the
expired list and once that new lease expired there would be duplicate IP
addresses in server->expired_list and it could keep growing.  Now when
we try to re-use a struct l_dhcp_lease in get_lease() we look up by IP
instead of by the MAC and we look at both server->lease_list and
server->expired_list.  This guarantees IPs are unique across the two
lists and their sizes are bounded by the pool size.