icmp6: Don't crash if client was stopped
It is possible that l_icmp6_client_stop is called re-entrantly from the
event notification callback (i.e. as a result of
L_ICMP6_CLIENT_EVENT_ROUTER_FOUND).  In this case the router
advertisement is freed and reset to NULL.  Detect this case and return
an -ECANCELED error to be handled accordingly.