dhcp6: Fix l_dhcp6_client_get_lease returning NULL
When a new lease is obtained, l_dhcp6_client fires off a LEASE_OBTAINED
event.  Unfortunately, the state of l_dhcp6_client object still remains
in SOLICITING/REQUESTING state, so calls to l_dhcp6_client_get_lease
return NULL.

Fix this by entering into the BOUND state when the lease is being setup
and prior to sending the event.  This allows clients who are interested
in the obtained lease to retrieve it properly.