base64: Null terminate encoding
l_base64_encode was returning a char* without any NULL terminator. This
appeared to be written to satisfy the only use in ELL (pem.c) but made
for a confusing public API. Anyone using this API would expect a char*
return to be NULL terminated.

Now l_base64_encode will NULL terminate which also removes the need for
the length out parameter.

pem.c was updated to use strlen rather than rely on the out parameter.