dhcp: handle multiple offers from the same server
It was found that some DHCP servers sometimes send multiple OFFERs.
ell's DHCP implementation always picks the first OFFER it receives and
uses it to form the REQUEST message.  On such servers the REQUEST is
never replied to, however the REQUEST to the subsequent OFFER works.

The RFC is a bit fuzzy on how multiple DHCP offers from the same server
should be treated, but in general it is up to the client implementation
to pick which OFFER to use.

If a new OFFER arrives from the same server ell sent a REQUEST to, the new
OFFER will be used to form a new REQUEST and the REQUESTING state will
be re-entered.