ecc: make l_ecc_point_from_data fully constant time
This was modified earlier to be constant time but there was one
minor problem with how 'sub' was determined. That logic could short
circuit which could alter the timing. The change in timing would be
extremely small, especially compared to the previous fix making
_vli_mod_sub() execute always but in any case there would be a small
difference.

This patch uses l_secure_select_byte to select either of the two
subtraction conditions which depend on the point type.