utf8: fixed bit shift larger than type size
The C standard requires the bit shift result value width to be greater
than or equal to the number of bits being shifted, otherwise the behavior
is undefined. In this case str[0] was being cast to an unsigned char
(1 byte) but being shifted 24 bits. This resulted, on some compilers,
with a runtime error:

runtime error: left shift of 255 by 24 places cannot be represented in type 'int'

Casting str[0] to an unsigned int satisfies the width requirements
for this bit shift.

Fixes: 9843abcd84bc ("utf8: cast to avoid left shift of negative value (ubsan)")