dhcp6: Set lease expiry based on frame reception times
Use the SO_TIMESTAMP socket option to request the packet rx timestamps
from the kernel and pass those values from dhcp6-transport to
dhcp6-client to use in lease expiry time calculation.

Apparently linux has a few different variants of SO_TIMESTAMP, use the
most basic one.  Additionally the rx times can be obtained without
SO_TIMESTAMP using the SIOCGSTAMP ioctl on each packet.  The advantage
would be that we'd only parse the timestamps for the frames where we
actually care about timing.  The disadvantage would be an extra syscall
for those frames, which we don't incur with SO_TIMESTAMP.