netconfig: Add netconfig API
This adds the netconfig class that will encapsulate static IP
configuration, DHCP, DHCP6, ICMP6, SLAAC logic and free the user of
manually handling DHCP/DHCP6 events to implement netconfig policies.
The user will still need to handle committing received DNS and mdns
information.

In this patch only IPv4 DHCP is implemented and no configurable settings
are added.

When events are emitted for address and/or route changes, the changes
are split into "added", "updated" and "removed" lists so that the user
can decide apply those changes without first removing all addresses and
routes from the interface and the re-adding the complete new list.
They can also decide to let another entity apply the changes.  The
obvious downside is the proliferation of lists.