cert-crypto: pragma false positive (-Wmaybe-uninitialized)
On musl-gcc 'bmpstring' was detected as maybe uninitialized. This
is actually a false positive since the usage is guarded by
if (p_len) but the compiler isn't aware. Pragma to ignore this
warning.

ell/cert-crypto.c: In function 'cert_pkcs12_pbkdf':
ell/cert-crypto.c:246:17: error: 'bmpstring' may be used uninitialized in this function [-Werror=maybe-uninitialized]
  246 |                 explicit_bzero(bmpstring, passwd_len);