netconfig: Ratelimit route lifetime updates
Optimize away prefix route updates when the expiry time is far enough
in the future that we expect to receive at least 10 Router
Advertisements before then.  For that purpose start keeping an extra
record for each route where we save the last time this route was
included in an RA and the maximum interval between two RAs with this
route seen so far.  By working with the maximum intervals we avoid
worrying about different RA content policies where some routes may be
skipped in some RAs and there even might be multiple routers sending
RAs with subsets of the same prefixes.

This new mechanism can be made configurable if desired but it's always
enabled for now.