netconfig: Set a limit on number of routes from ICMPv6
While some networks can be configured in a way that prevents rogue
Router Advertisements from reaching clients, add a very basic mechanism
to avoid a denial-of-service in such a case by limiting the number of
routes we track from RAs.  This won't prevent the loss of routing but
might prevent OOM crashes or slowdowns.