netconfig: Start ACD before committing static IPv4
When the IPv4 is configured to be static we can use ACD in order to
check that the IP is available and not already in use.  If a
conflict is found netconfig will be reset and no IP will be set on
the interface.  The ACD client is left with the default
'defend once' policy, and probes are not turned off.  This will
increase connection time, but for static IP it is considered the
safe approach.