useful: Add a cleanup handler for fd variables
Allow declaring an fd variable with a simple
    _auto_(close) int fd = -1;
to be able to skip the close() call.  This is useful for very simple
file accesses where it's not worth using l_io with
l_io_set_close_on_destroy.

As an example, update netconfig_proc_write_ipv6_setting to use
_auto_(close) where it helps in returning the errno from the read() call
(on error) which would otherwise be clobbered by the close() call.