icmp6: Switch socket from AF_INET6 to AF_PACKET
In order to be able to send our Router Solicitations before the
interface has a confirmed link-local address we have to be able to
supply our own IPv6 header and skip source address selection and routing
in the kernel.  It is at least tricky to do with an AF_INET6 raw socket
as some socket features, like SO_DONTROUTE/MSG_DONTROUTE, are not
supported and others don't suppress the source address selection/checks.
Additionally since we need to use the IPv6 unspecified address, which is
represented as all 0 bytes, some syscalls interpret it as no value
supplied.

Build/parse the IPv6 header on our own and send/receive using an
AF_PACKET socket.

Once we have the link-local address, we switch to sending from that
address.  Add l_icmp6_client_set_link_local_address for the user to give
it to us as they're likely to need it for DHCPv6.